fix: detect small QR codes in large images via crop+upscale
Previously, the upscale fallback for small QR codes only triggered when
the entire image was < 200px. A small QR code (e.g. 91x81px) in a large
image (e.g. 793x1121px scanned page) would never trigger upscaling,
causing detection to fail.

Now the scanner saves the bounding box of detected QR finder lines before
decoding. If the new `retry_undecoded_regions` config option is set it
will try to automatically crop+upscale.

+++ src/decoders/code128.rs
@@ -235,7 +235,7 @@ fn decode6(dcode: &ImageScanner) -> i8 {
         return -1;
     }
 
-    (c & 0x7f) as i8
+    c & 0x7f
 }
 
 /// Validate checksum
